Add Promocode

+++ ArmenianChairDogsitting.Data/Entities/Promocode.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Promo { get; set; }
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }    
     }
